feat: proxy 설정 > 리버트
