+++ package.json
@@ -18,7 +18,6 @@
 		"http-proxy-middleware": "^2.0.6",
 		"next": "^13.4.19",
 		"next-auth": "^4.23.1",
-		"next-http-proxy-middleware": "^1.2.5",
 		"next-images": "^1.8.5",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
